# docs/swagger.yaml
definitions:
  api_types.CalculationDetailedResponse:
    properties:
      completed_at:
        type: string
      created_at:
        type: string
      creator:
        $ref: '#/definitions/api_types.UserResponse'
      id:
        type: integer
      items:
        description: Главное отличие - наличие Items
        items:
          $ref: '#/definitions/api_types.CalculationItemResponse'
        type: array
      moderator:
        $ref: '#/definitions/api_types.UserResponse'
      status:
        type: string
      submitted_at:
        type: string
      total_profit:
        type: number
    type: object
  api_types.CalculationItemResponse:
    properties:
      coefficient:
        type: number
      final_dividend:
        description: omitempty скроет поле, если оно null
        type: number
      fine:
        type: number
      shareholder:
        $ref: '#/definitions/api_types.ShareholderResponse'
    type: object
...
  api_types.CalculationResponse:
    properties:
      completed_at:
        type: string
      created_at:
        type: string
      creator:
        $ref: '#/definitions/api_types.UserResponse'
      id:
        type: integer
      moderator:
        $ref: '#/definitions/api_types.UserResponse'
      status:
        type: string
      submitted_at:
        type: string
      total_profit:
        type: number
    type: object
  api_types.CalculationUpdateRequest:
    properties:
      total_profit:
        type: number
    type: object
  api_types.CartInfoResponse:
    properties:
      count:
        type: integer
      draft_id:
        type: integer
    type: object
  api_types.ModerationRequest:
    properties:
      status:
        description: '"completed" или "rejected"'
        type: string
    type: object
...
  api_types.ShareholderResponse:
    properties:
      description:
        type: string
      id:
        type: integer
      image_url:
        description: Используем *string для nullable
        type: string
      name:
        type: string
      share:
        type: number
    type: object
...
  api_types.UserResponse:
    properties:
      id:
        type: integer
      is_moderator:
        type: boolean
      login:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /dividend-calculations:
    get:
      description: Get all calculations (except draft/deleted). Users see their own,
        moderators see all.
      parameters:
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Start date for filtering (YYYY-MM-DD)
        in: query
        name: from_date
        type: string
      - description: End date for filtering (YYYY-MM-DD)
        in: query
        name: to_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api_types.CalculationResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get list of calculations
      tags:
      - calculations
  /dividend-calculations/{id}:
    delete:
      description: Logically delete a draft calculation.
      parameters:
      - description: Calculation ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden (not creator or not a draft)
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a draft calculation
      tags:
      - calculations
    get:
      description: Get details of a single calculation by its ID.
      parameters:
      - description: Calculation ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_types.CalculationDetailedResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a calculation by ID
      tags:
      - calculations
    put:
      consumes:
      - application/json
      description: Update fields of a draft calculation (e.g., total_profit).
      parameters:
      - description: Calculation ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api_types.CalculationUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_types.CalculationResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden (not creator or not a draft)
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a draft calculation
      tags:
      - calculations
  /dividend-calculations/{id}/moderate:
    put:
      consumes:
      - application/json
      description: Complete or reject a 'submitted' calculation. Requires moderator
        rights.
      parameters:
      - description: Calculation ID
        in: path
        name: id
        required: true
        type: integer
      - description: Moderation action
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api_types.ModerationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_types.CalculationDetailedResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden (not a moderator or not a submitted calculation)
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Moderate a calculation (Moderator only)
      tags:
      - calculations
...
  /dividend-calculations/{id}/submit:
    put:
      description: Change the status of a draft calculation to 'submitted'.
      parameters:
      - description: Calculation ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_types.CalculationResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden (not creator or not a draft)
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Submit a draft calculation
      tags:
      - calculations
  /dividend-calculations/cart:
    get:
      description: Get current draft ID and item count for the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_types.CartInfoResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get cart info
      tags:
      - calculations
...
swagger: "2.0"